chore(device): eliminated create device form duplication
