+++ apps/ui/src/pages/devices/index.vue
@@ -1,54 +1,17 @@
 <!-- eslint-disable vue/multi-word-component-names -->
 <script setup lang="ts">
-import { computed, onMounted, reactive, ref, watch } from 'vue';
+import { computed, onMounted, ref } from 'vue';
 import { useRoute } from 'vue-router';
-import { FRAME_SIZE_CODES, getFrameSizeCodesForFormatCode, type CreateFilmDeviceRequest, type FilmDevice } from '@frollz2/schema';
+import { type FilmDevice } from '@frollz2/schema';
 import { useDeviceStore } from '../../stores/devices.js';
 import { useReferenceStore } from '../../stores/reference.js';
-import { useUiFeedback } from '../../composables/useUiFeedback.js';
-import { createIdempotencyKey } from '../../composables/idempotency.js';
+import CreateDeviceDialog from '../../components/CreateDeviceDialog.vue';
 
 const route = useRoute();
 const deviceStore = useDeviceStore();
 const referenceStore = useReferenceStore();
-const feedback = useUiFeedback();
 const search = ref<string | null>('');
 const isCreateDialogOpen = ref(false);
-const isCreating = ref(false);
-const idempotencyKey = ref(createIdempotencyKey());
-
-const createForm = reactive({
-  deviceTypeCode: 'camera',
-  filmFormatId: null as number | null,
-  frameSize: 'full_frame' as CreateFilmDeviceRequest['frameSize'],
-  loadMode: 'direct' as 'direct' | 'interchangeable_back',
-  make: '',
-  model: '',
-  name: '',
-  system: '',
-  brand: '',
-  slotCount: 2,
-  holderTypeId: null as number | null
-});
-
-const selectedFilmFormatCode = computed(() => {
-  const selectedFilmFormat = referenceStore.filmFormats.find((format) => format.id === createForm.filmFormatId);
-  return selectedFilmFormat?.code ?? null;
-});
-
-const frameSizeOptions = computed(() => {
-  if (!selectedFilmFormatCode.value) {
-    return [...FRAME_SIZE_CODES];
-  }
-  return [...getFrameSizeCodesForFormatCode(selectedFilmFormatCode.value)];
-});
-
-const isFormFieldDisabled = computed(() => !createForm.filmFormatId || !createForm.deviceTypeCode);
-
-const isFrameSizeDisabled = computed(() =>
-  isFormFieldDisabled.value ||
-  (createForm.deviceTypeCode === 'camera' && createForm.loadMode !== 'direct')
-);
 
 const routeTypeFilter = computed(() => {
   const value = route.meta.deviceTypeFilter;
@@ -100,27 +63,6 @@ const columns = [
   }
 ];
 
-const deviceTypeOptions = computed(() =>
-  referenceStore.deviceTypes.map((type) => ({
-    label: type.label,
-    value: type.code
-  }))
-);
-
-const filmFormatOptions = computed(() =>
-  referenceStore.filmFormats.map((format) => ({
-    label: format.label,
-    value: format.id
-  }))
-);
-
-const holderTypeOptions = computed(() =>
-  referenceStore.holderTypes.map((type) => ({
-    label: type.label,
-    value: type.id
-  }))
-);
-
 function deviceLabel(device: FilmDevice): string {
   if (device.deviceTypeCode === 'camera') {
     return `${device.make} ${device.model}`;
@@ -131,112 +73,6 @@ function deviceLabel(device: FilmDevice): string {
   }
 
   return device.name;
-}
-
-function resetCreateForm(): void {
-  createForm.deviceTypeCode = 'camera';
-  createForm.filmFormatId = null;
-  createForm.frameSize = FRAME_SIZE_CODES[0];
-  createForm.loadMode = 'direct';
-  createForm.make = '';
-  createForm.model = '';
-  createForm.name = '';
-  createForm.system = '';
-  createForm.brand = '';
-  createForm.slotCount = 2;
-  createForm.holderTypeId = null;
-  idempotencyKey.value = createIdempotencyKey();
-}
-
-watch(frameSizeOptions, (options) => {
-  if (options.length === 0) {
-    return;
-  }
-  if (!options.includes(createForm.frameSize)) {
-    const nextFrameSize = options[0];
-    if (!nextFrameSize) {
-      return;
-    }
-    createForm.frameSize = nextFrameSize;
-  }
-}, { immediate: true });
-
-function openCreateDialog(): void {
-  resetCreateForm();
-  isCreateDialogOpen.value = true;
-}
-
-async function submitCreate(): Promise<void> {
-  if (isCreating.value) {
-    return;
-  }
-
-  const deviceType = referenceStore.deviceTypes.find((type) => type.code === createForm.deviceTypeCode);
-  if (!deviceType) {
-    feedback.error('Select a device type.');
-    return;
-  }
-
-  let payload: CreateFilmDeviceRequest;
-
-  if (createForm.deviceTypeCode === 'camera') {
-    if (!createForm.make.trim() || !createForm.model.trim() || !createForm.filmFormatId) {
-      feedback.error('Make, model, and format are required.');
-      return;
-    }
-
-    payload = {
-      deviceTypeCode: 'camera',
-      deviceTypeId: deviceType.id,
-      filmFormatId: createForm.filmFormatId,
-      frameSize: createForm.loadMode === 'direct' ? createForm.frameSize as CreateFilmDeviceRequest['frameSize'] : 'full_frame',
-      make: createForm.make.trim(),
-      model: createForm.model.trim(),
-      canUnload: true,
-      loadMode: createForm.loadMode
-    };
-  } else if (createForm.deviceTypeCode === 'interchangeable_back') {
-    if (!createForm.name.trim() || !createForm.system.trim() || !createForm.filmFormatId) {
-      feedback.error('Name, system, and format are required.');
-      return;
-    }
-
-    payload = {
-      deviceTypeCode: 'interchangeable_back',
-      deviceTypeId: deviceType.id,
-      filmFormatId: createForm.filmFormatId,
-      frameSize: createForm.frameSize as CreateFilmDeviceRequest['frameSize'],
-      name: createForm.name.trim(),
-      system: createForm.system.trim()
-    };
-  } else {
-    if (!createForm.name.trim() || !createForm.brand.trim() || !createForm.holderTypeId || !createForm.filmFormatId) {
-      feedback.error('Holder name, brand, holder type, and format are required.');
-      return;
-    }
-
-    payload = {
-      deviceTypeCode: 'film_holder',
-      deviceTypeId: deviceType.id,
-      filmFormatId: createForm.filmFormatId,
-      frameSize: createForm.frameSize as CreateFilmDeviceRequest['frameSize'],
-      name: createForm.name.trim(),
-      brand: createForm.brand.trim(),
-      slotCount: createForm.slotCount as 1 | 2,
-      holderTypeId: createForm.holderTypeId
-    };
-  }
-
-  isCreating.value = true;
-  try {
-    await deviceStore.createDevice(payload, idempotencyKey.value);
-    feedback.success('Device created.');
-    isCreateDialogOpen.value = false;
-  } catch (error) {
-    feedback.error(feedback.toErrorMessage(error, 'Failed to create device.'));
-  } finally {
-    isCreating.value = false;
-  }
 }
 
 onMounted(async () => {
@@ -252,7 +88,7 @@ onMounted(async () => {
         <div class="text-subtitle2 text-grey-7">Cameras, backs, and holders.</div>
       </div>
       <div class="row q-gutter-sm">
-        <q-btn color="primary" label="Add device" @click="openCreateDialog" />
+        <q-btn color="primary" label="Add device" @click="isCreateDialogOpen = true" />
         <q-btn flat color="primary" label="Refresh" @click="deviceStore.loadDevices" />
       </div>
     </div>
@@ -278,71 +114,9 @@ onMounted(async () => {
       </template>
     </q-table>
 
-    <q-dialog v-model="isCreateDialogOpen">
-      <q-card class="full-width">
-        <q-card-section>
-          <div class="text-h6">Create device</div>
-        </q-card-section>
-
-        <q-card-section>
-          <q-form class="column q-gutter-md" @submit="submitCreate">
-            <q-select
-              v-model="createForm.deviceTypeCode"
-              filled
-              emit-value
-              map-options
-              :options="deviceTypeOptions"
-              label="Device type"
-            />
-            <q-select
-              v-model="createForm.filmFormatId"
-              filled
-              emit-value
-              map-options
-              :options="filmFormatOptions"
-              label="Film format"
-            />
-            <q-select v-model="createForm.frameSize" filled :options="frameSizeOptions" label="Frame size" :disable="isFrameSizeDisabled" />
-
-            <template v-if="createForm.deviceTypeCode === 'camera'">
-              <q-toggle v-model="createForm.loadMode" true-value="direct" false-value="interchangeable_back" label="Is this camera directly loadable?" :disable="isFormFieldDisabled" />
-              <q-input v-model="createForm.make" filled label="Make" :disable="isFormFieldDisabled" />
-              <q-input v-model="createForm.model" filled label="Model" :disable="isFormFieldDisabled" />
-            </template>
-
-            <template v-else-if="createForm.deviceTypeCode === 'interchangeable_back'">
-              <q-input v-model="createForm.name" filled label="Name" :disable="isFormFieldDisabled" />
-              <q-input v-model="createForm.system" filled label="System" :disable="isFormFieldDisabled" />
-            </template>
-
-            <template v-else>
-              <q-input v-model="createForm.name" filled label="Holder name" :disable="isFormFieldDisabled" />
-              <q-input v-model="createForm.brand" filled label="Brand" :disable="isFormFieldDisabled" />
-              <q-select
-                v-model="createForm.slotCount"
-                filled
-                :options="[1, 2]"
-                label="Slot count"
-                :disable="isFormFieldDisabled"
-              />
-              <q-select
-                v-model="createForm.holderTypeId"
-                filled
-                emit-value
-                map-options
-                :options="holderTypeOptions"
-                label="Holder type"
-                :disable="isFormFieldDisabled"
-              />
-            </template>
-          </q-form>
-        </q-card-section>
-
-        <q-card-actions align="right">
-          <q-btn v-close-popup flat label="Cancel" />
-          <q-btn color="primary" label="Create" :loading="isCreating" @click="submitCreate" />
-        </q-card-actions>
-      </q-card>
-    </q-dialog>
+    <CreateDeviceDialog
+      v-model="isCreateDialogOpen"
+      :device-type-filter="routeTypeFilter ?? undefined"
+    />
   </q-page>
 </template>
